requirements.txt
Markdown==3.3.3
pylint==2.7.0
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
requests==2.25.1
urllib3==1.26.3
websocket-client==0.56
zipp==3.4.0